requirements.txt
google-genai>=0.1.0
python-telegram-bot>=20.0
python-docx>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
